backend/tests/conftest.py
import pytest
from api import create_app
from api.config.config import TestingConfig
from api.extensions import db


@pytest.fixture
def app():
    app = create_app(TestingConfig)

    with app.app_context():
        db.create_all()

    yield app

    with app.app_context():
        db.drop_all()


@pytest.fixture
def client(app):
    return app.test_client()