backend/requirements.txt
#backend
flask==3.1.3
flask-bcrypt==1.0.1
flask-jwt-extended==4.7.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-caching==2.3.1
flasgger==0.9.7.1
sqlalchemy==2.0.48
alembic==1.18.4
psycopg2==2.9.11
pymysql==1.1.1
python-dotenv==1.2.2
pandas==3.0.2
numpy==2.4.4
openpyxl==3.1.5
unidecode==1.4.0
cryptography==46.0.5
paramiko==3.5.1
sshtunnel==0.4.0